requirements.txt
pandas
python-binance
asyncio
sqlalchemy
nest_asyncio